external/webkit/LayoutTests/platform/android/webgl-conformance/dumpAsText/conformance/glsl/implicit/assign_int_to_float.vert-expected.txt

CONSOLE MESSAGE: line 24: *** Error compiling shader '[object WebGLShader]':ERROR: 0:4: '=' :  cannot convert from 'const mediump int' to 'float'

implicit cast assing int to float should fail

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

test: implicit cast assing int to float should fail
test vertex shader
default fragment shaderPASS implicit cast assing int to float should fail
PASS successfullyParsed is true

TEST COMPLETE
